Now fixes float64 int types
